+++ private/mdToMdAst.js
@@ -3,6 +3,7 @@
 const gfm = require('remark-gfm');
 const parse = require('remark-parse');
 const unified = require('unified');
+const removePositionData = require('unist-util-remove-position');
 const replaceJsdocLinks = require('./replaceJsdocLinks');
 const unescapeJsdoc = require('./unescapeJsdoc');
 
@@ -19,8 +20,20 @@ module.exports = function mdToMdAst(markdown, members) {
   if (typeof markdown !== 'string')
     throw new TypeError('First argument “markdown” must be a string.');
 
-  return unified()
-    .use(parse)
-    .use(gfm)
-    .parse(replaceJsdocLinks(unescapeJsdoc(markdown), members)).children;
+  // The AST nodes from a parsed markdown string contain `position` data
+  // (https://github.com/syntax-tree/unist#position). This data should be
+  // removed because it will no longer be correct once these AST nodes are
+  // inserted into another AST. While leaving the incorrect data in place is
+  // technically more efficient and harmless to the public API, it bloats
+  // private test snapshots.
+  return removePositionData(
+    unified()
+      .use(parse)
+      .use(gfm)
+      .parse(replaceJsdocLinks(unescapeJsdoc(markdown), members)),
+
+    // Delete the `position` properties from nodes instead of only replacing
+    // their values with `undefined`.
+    true
+  ).children;
 };